requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
bleach==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
chromedriver-installer==0.0.6
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
decorator==4.3.0
entrypoints==0.2.3
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipykernel==4.8.2
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.2.3
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.0
pandocfilters==1.4.2
parsel==1.5.0
pexpect==4.6.0
pickleshare==0.7.4
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
pymongo==3.7.1
pyOpenSSL==18.0.0
python-dateutil==2.7.3
pyzmq==17.1.2
qtconsole==4.4.0
queuelib==1.5.0
requests==2.19.1
Scrapy==1.5.1
scrapy-mongodb==0.12.0
scrapy-selenium==0.0.3
selenium==3.13.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
Twisted==18.7.0
urllib3==1.23
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
zope.interface==4.5.0
dateparser==0.7.0
arrow==0.12.1
html2text==2018.1.9
beautifulsoup4==4.6.3
botocore
boto3==1.8.1
parsel==1.5.0
scrapy_proxies==0.4
firebase-admin==2.13.0
langdetect==1.0.7